refactor: 클래스명 CardSymbol -> Suit, CardScore -> Rank로 수정
