+++ src/test/java/blackjack/view/RankNameTest.java
@@ -1,6 +1,6 @@
 package blackjack.view;
 
-import blackjack.domain.card.CardSymbol;
+import blackjack.domain.card.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -8,15 +8,15 @@ import java.util.Arrays;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-@DisplayName("카드 문양 이름")
-class CardSymbolNameTest {
+@DisplayName("카드 점수 이름")
+class RankNameTest {
     @Test
     @DisplayName("모두 변환이 된다.")
     void convertAllTest() {
         // given & when & then
         assertThatCode(() ->
-                Arrays.stream(CardSymbol.values())
-                        .forEach(CardSymbolName::convert)
+                Arrays.stream(Rank.values())
+                        .forEach(CardScoreName::convert)
         )
                 .doesNotThrowAnyException();
     }
