+++ src/test/java/blackjack/view/SuitNameTest.java
@@ -1,6 +1,6 @@
 package blackjack.view;
 
-import blackjack.domain.card.CardScore;
+import blackjack.domain.card.Suit;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -8,15 +8,15 @@ import java.util.Arrays;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-@DisplayName("카드 점수 이름")
-class CardScoreNameTest {
+@DisplayName("카드 문양 이름")
+class SuitNameTest {
     @Test
     @DisplayName("모두 변환이 된다.")
     void convertAllTest() {
         // given & when & then
         assertThatCode(() ->
-                Arrays.stream(CardScore.values())
-                        .forEach(CardScoreName::convert)
+                Arrays.stream(Suit.values())
+                        .forEach(CardSymbolName::convert)
         )
                 .doesNotThrowAnyException();
     }
